CURB-3197 Improved favorites notes and quantity section on browser layout
